backend/requirements.txt
flask
mysql-connector-python
opencv-python
face-recognition
tensorflow
numpy
pillow
scikit-learn